:sparkles: Bad request excpetion handler
400 exception handler and some naming changes

+++ service/document_service.py
@@ -1,3 +1,4 @@
+import re
 from html.parser import HTMLParser
 
 from fastapi import HTTPException
@@ -14,11 +15,16 @@ from utils.env_variables_reader import EnvVariablesReader
 import uuid
 from bs4 import BeautifulSoup
 from dto.exceptions.not_found_exception import NotFoundException
+from dto.exceptions.bad_request_exception import BadRequestException
+from dto.response.simple_url_response import SimpleUrlResponse
 
 templates = Jinja2Templates(directory="templates")
 
 
 class DocumentService:
+    guid_pattern = "^(?:\\{{0,1}(?:[0-9a-fA-F]){8}-(?:[0-9a-fA-F]){4}-(?:[0-9a-fA-F]){4}-(?:[0-9a-fA-F]){4}-(?:[" \
+                   "0-9a-fA-F]){12}\\}{0,1})$"
+
     async def create_document(self, db: Session, request: Request, data: document_request.DocumentRequest):
         template = templates.TemplateResponse("document.html", {"request": request, "content": data.content})
         soup = BeautifulSoup(template.body)
@@ -27,9 +33,11 @@ class DocumentService:
         db.add(document)
         db.commit()
         db.refresh(document)
-        return SimpleMessageResponse(f"{app_base_url}/documents/{document.guid}")
+        return SimpleUrlResponse(f"{app_base_url}/documents/{document.guid}")
 
     async def get_document(self, db: Session, guid: str):
+        if re.match(self.guid_pattern, guid) is None:
+            raise BadRequestException()
         document = db.query(Document).filter(Document.guid == uuid.UUID(guid)).first()
         if document is not None:
             res = HTMLResponse(document.content).body
@@ -39,6 +47,8 @@ class DocumentService:
             raise NotFoundException("document")
 
     async def delete(self, db: Session, guid: str):
+        if re.match(self.guid_pattern, guid) is None:
+            raise BadRequestException()
         document = db.query(Document).filter(Document.guid == uuid.UUID(guid)).first()
         if document is not None:
             db.query(Document).filter(Document.guid == uuid.UUID(guid)).delete()
